get rid of redundant "Input:", "Output:" specs
Every encoding in Section 5 had *exactly* the same text specifying
"Input: alpha..." and "Output: (x, y)...". There is already text
in Section 5.1 specifying the generic i/o interface, so all of this
text is redundant. Keeping to the particulars of a given map makes
it easier to understand the differences among them.