pass over relwork
This commit passes over the relwork, making it a bit more compact
and improving coverage of relwork, including map-to-bitstring.
Closes #113.

This commit also removes the Try-and-Increment method, which is
redundant with the description of MapToGroup from BLS01 in relwork.

Since try-and-increment is not recommended for any curve, there is
no need to provide pseudocode; doing so may accidentally encourage
its use. Interested readers can follow the BLS01 cite.