injective -> nonuniform
Thinking about it more, 'injective' is not a great choice for
non-RO encodings, because in fact they are definitely *not*
injective! In particular, since hash2base takes bitstrings
of arbitrary length and outputs elements of F, there must
exist colliding inputs by a simple counting argument (though
presumably they are infeasible to find). This immediately means
that any encoding that uses hash2base is not injective, because
there exist two inputs that produce the same output.