change variables for elliptic curve equations
This commit is intended to help readers distinguish among the four
elliptic curve equations used in this document:

- Short Weierstrass, y^2 = x^3 + A * x + B

- Long Weierstrass, Y^2 = X^3 + C * X^2 + D * X

- Montgomery, K * W^2 = V^3 + J * V^2 + V

- Twisted Edwards, a * v^2 + w^2 = 1 + d * v^2 * w^2

The map "Elligator 2, A == 0" is now called "Elligator 2, C == 0"
for consistency with the above changes.

I have not changed the example code to use different variable names,
because I don't think it makes a difference for clarity.

I chose the above variable names to avoid clashing with existing
notation. I'll admit they're not necessarily the most beautiful
choices, but very little else is available in the ASCII charset
and these seem to be the best choices.