sqrt_ratio constant S can be set to Z
Define l, o such that q - 1 = 2^l * o and o is odd.
The requirement for correctness of the Tonelli-Shanks
algorithm is just that S^o generates the multiplicative
subgroup of F with order 2^l.

When S is non-square, this is guaranteed to be the case:
S^o must also be non-square because o is odd, and it must
be an element of the order-2^l subgroup by the definition
of o. All non-square elements of the order-2^l subgroup
are generators, and vice-versa.